app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
gunicorn
psycopg2-binary
python-multipart